requirements.txt
tensorflow==2.17.0
keras==3.5.0
streamlit
numpy
Pillow
flask
beautifulsoup4
requests